added status codes

+++ src/main/java/bo/usfx/springneuroapi/controller/NeurodiversityController.java
@@ -33,25 +33,30 @@ public final class NeurodiversityController {
     }
 
     // get entity by method getId
+
     @GetMapping("/api/v1/neurodiversity/{id}")
     public ResponseEntity<?> getById(@PathVariable final String id) {
+
         var neuro = neurodivergencyRepository.findById(id);
-        if (neuro != null) {
-            return ResponseEntity.ok(neuro);
+
+        if (neuro.isPresent()) {
+            return new ResponseEntity<>(neuro, HttpStatus.OK);
         }
-        return null;
+        return new ResponseEntity<>("Document with id: " + id + " was not found", HttpStatus.NOT_FOUND);
+
     }
 
     // get entity by method getName
     @GetMapping("/api/v1/neurodiversity-name/{name}")
     public ResponseEntity<?> getByName(@PathVariable final String name) {
         var neuro = neurodivergencyRepository.findByName(name);
-        if (neuro != null) {
+        if (!neuro.isEmpty()) {
             return ResponseEntity.ok(neuro);
         }
-        return null;
+        return new ResponseEntity<>("Document with name: " + name + " was not found", HttpStatus.NOT_FOUND);
     }
 
+    // Post Request
     @PostMapping("/api/v1/neurodiversity")
     public ResponseEntity<?> createEntity(@RequestBody final Neurodiversity neurodiversity) {
         neurodivergencyRepository.save(neurodiversity);
@@ -61,7 +66,7 @@ public final class NeurodiversityController {
     // Put Request //
     @PutMapping("/api/v1/neurodiversity/{id}")
     public ResponseEntity<Neurodiversity> updateN(@PathVariable(value = "id") final String id,
-            @RequestBody final Map<String, Object> fields) {
+                                                  @RequestBody final Map<String, Object> fields) {
         Neurodiversity updatedNeurodiversity = updateNeurodiversityFields(id, fields);
         neurodivergencyRepository.save(updatedNeurodiversity);
         return ResponseEntity.ok(updatedNeurodiversity);
@@ -84,12 +89,14 @@ public final class NeurodiversityController {
     // Delete Request
     @DeleteMapping("/api/v1/neurodiversity/{id}")
     public ResponseEntity<?> deleteNeuro(@PathVariable("id") final String id) {
-        try {
+
+        var neuroObjec = neurodivergencyRepository.findById(id);
+        if (neuroObjec.isPresent()) {
             neurodivergencyRepository.deleteById(id);
             return new ResponseEntity<>("Successfully deleted", HttpStatus.OK);
-        } catch (Exception e) {
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
         }
+        return new ResponseEntity<>("The document does not exist", HttpStatus.NOT_FOUND);
+
     }
 
     // @RequestParams //
